moved widgets added changes
