+++ widgetastic_manageiq/__init__.py
@@ -4720,20 +4720,3 @@ class HiddenFileInput(BaseFileInput):
     def is_displayed(self):
         self.browser.set_attribute("style", "display;", self)
         return self.browser.is_displayed(self)
-
-
-class MigrationDropdown(Dropdown):
-    """Represents the migration plan dropdown of v2v.
-
-    Args:
-        text: Text of the button, can be inner text or the title attribute.
-    """
-    ROOT = ParametrizedLocator('.//div[contains(@class, "dropdown") and '
-                               './/button[contains(@id, "dropdown-filter")]]')
-    BUTTON_LOCATOR = './/*[contains(@id, "dropdown-filter")]'
-    ITEMS_LOCATOR = './/ul[contains(@aria-labelledby,"dropdown-filter")]/li/a'
-    ITEM_LOCATOR = './/ul[contains(@aria-labelledby,"dropdown-filter")]/li/a[normalize-space(.)={}]'
-
-    def __init__(self, parent, text, logger=None):
-        Widget.__init__(self, parent, logger=logger)
-        self.text = text
